1. How git pull works under the hood? -----------------------------------
Under the hood git pull is actually running two separate commands:
git fetch which checks for any changes in the remote since the last fetch,
and either git merge or git rebase , depending on how the command is configured,
which then merges/rebases these changes into your local branch.

2. Understanding Git SHA
----------------------
This SHA is calculated based on the contents of the object using a cryptographic hash function.
The SHA ensures data integrity and guarantees that the object remains unchanged.
Any modification to the content results in a different SHA, alerting Git to the change.